Move _tool_call_writer ContextVar next to its writer
Relocates the per-tool-call writer ContextVar from `langgraph.config`
to `langgraph.pregel._tools`, where `StreamToolCallHandler` (the only
producer) already lives. `ToolRuntime.emit_output_delta` (the only
consumer) imports it from there directly.

`langgraph.config` is the public surface for `get_config` /
`get_store` / `get_stream_writer`; an internal ContextVar shared
between the tool handler and `ToolRuntime` doesn't belong there.
